% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameters.R
\name{check_parameters}
\alias{check_parameters}
\title{Validate parameters without knitting a Rmd file}
\usage{
check_parameters(
  rmarkdown_document,
  params,
  allow_NULL = FALSE,
  allow_NA = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{rmarkdown_document}{path to a Rmarkdown document (i.e. \code{*.Rmd}) to be inspected.}

\item{params}{a named list of arguments to use for the report. Names must be a subset of the declarations in the yaml header of Rmarkdown report supplied.}

\item{allow_NULL}{Should \code{NULL} as a parameter value be accepted for parameters? This takes precedence over how check functions may handle \code{NULL} values.}

\item{allow_NA}{Should \code{NA} as a parameter value be accepted for parameters? This takes precedence over how check functions may handle \code{NA} values.}

\item{verbose}{Set \code{TRUE} to get verbose messages about the checks performed.}
}
\value{
The function returns \code{TRUE} if the supplied parameters are a subset of the parameters in the yaml header, and all checks were ok \code{TRUE}. In all other cases, it returns \code{FALSE}. Use \code{verbose=TRUE} to get information about what went wrong.
}
\description{
Validate parameters without knitting a Rmd file
}
\seealso{
\code{vignette("writing_checks", package = "rmarkdown.checks")} for a guide on how to declare check functions in the yaml header that work with this package.
}
